Adjust max entries and calculator size to fit 20 digits
For solutions larger than 20 digits, javascript starts using e to
to represent the answer. By setting the max digits for an entry to 20
and allowing space for all 20, it allows entering 9 for all 20 digits
and adding 1 to provide and show the answer of 1e+21 that still fits in
the display window.